Moved congestionCharges.ts into the common folder
